backend/requirements.txt
# SmartATS Backend Dependencies
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Document Processing
pdfplumber==0.10.4
python-docx==1.1.0

# NLP & Machine Learning
spacy==3.7.4
scikit-learn==1.4.0
numpy==1.26.4

# Utilities
pydantic==2.6.1
python-dotenv==1.0.1

# Optional: OpenAI for AI-powered suggestions
# openai==1.12.0
